Create third-culture-issues
There has been no issues over the past 10 years. The "Out Sheep" initiative was something that existed, but that didnt really impacted us. Of course, this initiative and the growing nationalism all over Europe and to be frank, all over the world is concerming. All democratic forces have to mobalize to stop this. Reacently, I become a victum of a verbal rasistic attack. Riding our bike back from school, the driver of a camper van almost hit. She was angry about a person driving "too slowly" in front of her that she forgot that she was in the middle of overtaking a bike (us) and came closer and closer. When I hadnt had space between her and the kerb to the foodpath, i started screeming to get her attention. She realised her mistake and made room for us again. But her huspend on the passenger seat opend the window and said in frensh, that if I dont speak frensh, I should leave the country. This wasnt so much of a shok for me as I am aware that this was a shock situation for him as well, but that created the feeling of "we are not wanted here" in my daughter. She cried and didnt know how to deal with the situation. For me this revieled a few things:

1. Rassim is real - the guy might have done this in an affect, but it shows his real belives. Believes the right-wing parties seed and exploid. In times of social media more then ever. 
2. The reality of being "third culture", is something we have to address more clearly with our kids. It is important to be aware of why we are here and who we are. This is especially important as we have seen from a daughter of our friends, what can happen, if this is contiously or uncontiously not addressed [Case study here]. And it also highlights:
3. How sensitive kids are for any kind of hostility and aggression. 

All of the above make me believe, we need to make sure to address this at home and in the school context. It is a reality, we have to be contious about and help our kids to navigate.